# backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TurPoint API",
    "version": "0.3.0",
    "description": "Tour marketplace API - operators, tours, group formation pricing, bookings, reviews, deals, and the smart trip planner. Most write operations and anything scoped to \"the logged-in account\" require a Bearer JWT; identity is always derived from the verified token, never from the request body. Sprint 3: all responses carry standard security headers (helmet); /api/auth/login and /api/auth/signup are rate-limited (20 requests/15min per IP) and every other route shares a broader 300 requests/15min ceiling; in production CORS is restricted to the configured frontend origin via CORS_ORIGIN."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/group-formations": {
      "get": {
        "summary": "Read-only lookup of a tour's current group state - does not create anything, safe to call just to preview status/price",
        "parameters": [
          {
            "name": "tour_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The tour's open (waiting/forming) group if one exists, otherwise its most recent confirmed group, otherwise null"
          },
          "400": {
            "description": "tour_id query param is required"
          }
        }
      }
    },
    "/group-formations/{id}": {
      "get": {
        "summary": "Get one group formation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group formation with live price_per_person"
          }
        }
      }
    },
    "/group-formations/expire-past-due": {
      "post": {
        "summary": "Cancel any waiting/forming groups whose tour date has passed, and cancel any bookings still pending on them. Meant to be called by a scheduler, not end users.",
        "parameters": [
          {
            "name": "x-cron-secret",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Required only if the server has CRON_SECRET set (production). Must match its value. Left open in local dev where CRON_SECRET is unset."
          }
        ],
        "responses": {
          "200": {
            "description": "List of cancelled group ids and cancelled booking ids"
          },
          "401": {
            "description": "CRON_SECRET is set on the server and the x-cron-secret header is missing or doesn't match"
          },
          "503": {
            "description": "Production only: CRON_SECRET is not configured on the server, so the route refuses to run"
          }
        }
      }
    },
    "/bookings": {
      "post": {
        "summary": "Book a tour (simulated payment) - requires login; identity comes from the token, not the request body. Charged at the tour's listed price, or the discounted price while a last-minute deal is active. Tours whose date has passed can't be booked.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "tour_id": 5,
                "seats": 2,
                "payment": {
                  "card_number": "4242424242424242"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking + ticket_code"
          },
          "400": {
            "description": "Missing tour_id, seats, or payment; seats not a positive whole number; seats exceeds the tour's max_participants; or the tour has already taken place"
          },
          "401": {
            "description": "Missing/invalid token"
          },
          "409": {
            "description": "Not enough spots left in the group"
          }
        }
      }
    },
    "/bookings/mine": {
      "get": {
        "summary": "An operator's bookings across ALL of their tours - for the operator dashboard",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of bookings, each with tour_title, tour_date, traveler_name, traveler_email"
          },
          "403": {
            "description": "This account has no operator profile"
          }
        }
      }
    },
    "/bookings/my-trips": {
      "get": {
        "summary": "The logged-in traveler's own bookings, across every tour they've booked - distinct from /bookings/mine, which is operator-scoped",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of bookings, each with tour_title, tour_date, tour_location"
          }
        }
      }
    },
    "/bookings/{id}": {
      "get": {
        "summary": "Get one booking / e-ticket - must be the traveler who made it",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking with nested tour info (title, date, location, route, operator_name)"
          },
          "403": {
            "description": "Not your booking"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/bookings/{id}/cancel": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Cancel a booking",
        "description": "Allowed for the traveler who made it or the operator who owns the tour. Only pending/confirmed bookings on tours that haven't happened yet. Frees the seats on the group.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The cancelled booking"
          },
          "400": {
            "description": "Tour already taken place"
          },
          "403": {
            "description": "Not your booking / tour"
          },
          "404": {
            "description": "Booking not found"
          },
          "409": {
            "description": "Already cancelled"
          }
        }
      }
    }
  }
}